Allow error in finder pattern locating
Sometimes our finders are not a perfect 1-1-3-1-1, but more like a
1-1.1-3.3-1-0.9. Switch our comparisons to check for a ratio of each
element compared to the first one. If they are close to 1 than we call
it a match

Also adjust the width of the finder to match the start -> end of the
entire RLE section. This reduces error as well