Fix linter script
Before all we could do was visualize. In the case of our linter script,
we tried to output to /dev/null because we just wanted to check that we
ran successfully. This stopped working as /dev/null is not a directory,
and we wanted to output multiple files

Now that we can do something that isn't visualize (decode the data),
just disable the visualization code while linting. This isn't optimal as
we may miss some problems in the visualization path, but it's good
enough for now