Update data iteration logic
Previously our iteration fell over once we hit the left finder. It turns
out that the simple skip/zig zag/turnaround loop was not sufficient.
We have another action of continuing in a straight line (when we hit the
left timing edge), and if we are in a finder we have to decide if it
means we should skip or turn around.

Since it gets pretty intertwined, merge the shouldSkip and
shouldTurnAround functions into one larger checkNextAction function.
This will determine what we should do next.