Fix extremely slow visualization time
When rendering the data walk order into the overlay svg, we were using
an allocPrint to convert the number into a string for printing. This
resulted in a very slow allocation from the GeneralPurposeAllocator in
what is effectively a hot loop. Replace with a bufprint which brings
rendering back to "instant enough"