Fix module size calculation on large QR codes
Our previous implementation used the finder sizes to guess the module
size. Since we know that the finders are 7 modules wide/tall, we should
know the size of a module once we find them. If the QR code is large
enough, any errors in this calculation are amplified as we move farther
from the top left corner.

Use the timing row/col to adjust for this. We have a good initial guess
which lets us find the timing row/col, but we use the timing data to
adjust and find an actual element size that stays consistent throughout
the whole image