Prepare for second binary
Right now we have a single binary used to decode and annotate a QR code.
As we develop more, more debug utilities will come into play. These
utilities may come with their own dependencies, so it makes sense to
put them in subfolders

With the way zig's module system works, we cannot import items from
paths that are not children or siblings of the module we are working in.
In order to do that we have to use zig modules

* Create a zig module "libqr" that has all of the code related to
  actually decoding the QR code
* Rename the "qr-decoder" binary to "qr-annotator", move xml builder and
  svg rendering code into a qr_annotator folder along with main
* Hack up the build script to support the new layout
  * Each module/binary is now defined by it's own function
  * Each function registers itself with zig build test
  * Removed a bunch of useless comments that were put there by zig
    init-exe